+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 The project follows Semantic Versioning (http://semver.org/)
 
+## 4.3.0 - tbd
+### Added
+- event listener for failed slmQueue jobs that send an email to all queue admins
+
+## 4.2.0 - 2017-11-17
+### Added
+- shutdown the slmQueue worker when the entityManager was closed (due to an
+  exception) so it can be restarted by supervisor with a fresh instance
+
 ## 4.1.0 - 2017-10-10
 ### Added
 - test & cs config